% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted,penfa-method}
\alias{fitted,penfa-method}
\title{Model-implied moments for a \code{penfa} object}
\usage{
\S4method{fitted}{penfa}(object, labels = TRUE)
}
\arguments{
\item{object}{An object of class \code{penfa}, found as a result of a call
to \code{penfa}.}

\item{labels}{Logical. If \code{TRUE}, the model-implied moments are named
according to the item names used in the model syntax.}
}
\description{
An S4 method returning the model-implied moments for an object of class
\code{penfa}. For every group, a list with the model-implied  moments is returned:
\code{cov} contains the implied covariance matrix, and \code{mean} the implied
mean vector. If just the covariance matrix is analyzed, only the \code{cov}
argument is returned.
}
\examples{

## --- Continuing the Examples from penfa manual page:
\dontshow{require(utils)
example("penfa", echo = FALSE)}
fitted(alasso_fit)
fitted(alasso_fitMG)


}
\seealso{
\code{\link{penfa}}, \code{\link{penfa-class}}
}
