% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,penfa-method}
\alias{show,penfa-method}
\title{Display a \code{penfa} object}
\usage{
\S4method{show}{penfa}(object)
}
\arguments{
\item{object}{An object of class \code{penfa}, found as a result of a call
to \code{penfa}.}
}
\description{
An S4 method printing a short summary of the estimation process, including
the optimization method, the specified penalty functions, the convergence
status, the number of iterations, the tuning selection strategy, and the
effective degrees of freedom.
}
\examples{

## --- Continuing the Examples from penfa manual page:
\dontshow{require(utils)
example("penfa", echo = FALSE)}
alasso_fit
alasso_fitMG


}
\seealso{
\code{\link{penfa}}, \code{\link{penfa-class}}
}
