% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,penfaData-method}
\alias{show,penfaData-method}
\title{Display details on the input data}
\usage{
\S4method{show}{penfaData}(object)
}
\arguments{
\item{object}{An object of class \code{penfaData}, found in the \code{Data}
slot from a \code{penfa} class object.}
}
\value{
An object reporting a short summary of the input data.
}
\description{
An S4 method showing information on the input data, including the number of
observations. In case of a multiple-group analysis, the sample sizes for each
group are displayed.
}
\seealso{
\code{\link{penfaData-class}}
}
