% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,penfaPenalty-method}
\alias{show,penfaPenalty-method}
\title{Display details on the penalization}
\usage{
\S4method{show}{penfaPenalty}(object)
}
\arguments{
\item{object}{An object of class \code{penfaPenalty}, found in the
\code{Penalize} slot from an object of \code{penfa} class.}
}
\description{
An S4 method showing information on the penalization process, including the
employed penalty functions and the model matrices they affect. Additionally,
it reports the optimal values of the tuning parameters and the tuning
parameter selection strategy. If the automatic procedure was used, the output
would also show the value of the influence factor, and the number of two-steps
iterations.
}
\seealso{
\code{\link{penfaPenalty-class}}
}
