% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef,penfa-method}
\alias{coef,penfa-method}
\title{Coefficients from a \code{penfa} object}
\usage{
\S4method{coef}{penfa}(object, type = "free", labels = TRUE)
}
\arguments{
\item{object}{An object of class \code{penfa}, found as a result of a call
to \code{penfa}.}

\item{type}{Character. If \code{type="free"}, only the estimated parameters
(both penalized and unpenalized) are returned. If \code{type="user"}, all
parameters listed in the parameter table are returned, including fixed
parameters.}

\item{labels}{Logical. If \code{TRUE}, parameters are returned with their
names.}
}
\description{
An S4 method returning the estimates of the model parameters
}
\examples{

## --- Continuing the Examples from penfa manual page:
\dontshow{require(utils)
example("penfa", echo = FALSE)}
coef(alasso_fit)



}
\seealso{
\code{\link{penfa}}, \code{\link{penfa-class}}
}
